[MicroWin] Add custom user settings

+++ functions/public/Invoke-WPFMicrowin.ps1
@@ -178,7 +178,22 @@ public class PowerManagement {
         Write-Host "Removal complete!"
 
         Write-Host "Create unattend.xml"
-        New-Unattend
+        #New-Unattend
+        if ($sync.MicrowinUserName.Text -eq "")
+        {
+            New-Unattend -userName "User"
+        }
+        else
+        {
+            if ($sync.MicrowinUserPassword.Password -eq "")
+            {
+                New-Unattend -userName "$($sync.MicrowinUserName.Text)"
+            }
+            else
+            {
+                New-Unattend -userName "$($sync.MicrowinUserName.Text)" -userPassword "$($sync.MicrowinUserPassword.Password)"
+            }
+        }
         Write-Host "Done Create unattend.xml"
         Write-Host "Copy unattend.xml file into the ISO"
         New-Item -ItemType Directory -Force -Path "$($scratchDir)\Windows\Panther"
@@ -248,7 +263,10 @@ public class PowerManagement {
             # reg delete "HKLM\zSOFTWARE\Microsoft\Windows\CurrentVersion\WindowsUpdate\Orchestrator\UScheduler\$_" /f | Out-Null
 
             # When in Offline Image
-            reg delete "HKLM\zSOFTWARE\Microsoft\WindowsUpdate\Orchestrator\UScheduler_Oobe\$_" /f | Out-Null
+            if (Test-Path "HKLM\zSOFTWARE\Microsoft\WindowsUpdate\Orchestrator\UScheduler_Oobe\$_")
+            {
+                reg delete "HKLM\zSOFTWARE\Microsoft\WindowsUpdate\Orchestrator\UScheduler_Oobe\$_" /f | Out-Null
+            }
         }
 
         reg add "HKLM\zSOFTWARE\Microsoft\Windows\CurrentVersion\Search" /v "SearchboxTaskbarMode" /t REG_DWORD /d 0 /f
